requirements.txt
hydra-core==1.1.0
numpy==1.23.1
protobuf==3.20.*
pytorch-lightning==1.5.10
torch==1.10.1
torchmetrics==0.7.3
torchvision==0.11.3
wandb==0.15.9
